fonts_rs_generator

Module generator

Generic code generators for build-time artifact generation.

Provides a unified GlyphGenerator trait and implementations for generating Rust constants and phf::Map codepoint tables from glyph metadata.

Re-exports

pub use codemap::CodemapGenerator;

pub use codemap::CodemapNaming;

pub use codemap::DefaultNaming;

pub use error::GenerateError;

pub use generate::GlyphGenerator;

pub use metadata_generator::MetadataGenerator;

pub use metadata_generator::escape_str;

pub use rust_constants::RustConstantsGenerator;

Modules

codemap

Generator for phf::Map codepoint lookup tables.

error

Error type for code generators.

generate

Trait for code generators that produce output from glyph metadata.

metadata_generator

Generic metadata generator trait for build-time phf::Map code generation.

rust_constants

Generator for Rust glyph name constants.